requirements.txt
pytest
alembic==1.3.1
amqp==2.5.2
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.1.0
black==19.10b0
blinker==1.4
celery==4.3.0
Click==7.0
configparser==4.0.2
Flask==1.1.1
flask-expects-json==1.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
kombu==4.6.6
linecache2==1.0.0
Mako==1.1.0
MarkupSafe==1.1.1
migrate==0.3.8
more-itertools==8.0.0
mysqlclient==1.4.6
pathspec==0.6.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
redis==3.3.11
regex==2019.11.1
six==1.13.0
SQLAlchemy==1.3.11
toml==0.10.0
traceback2==1.4.0
typed-ast==1.4.0
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
Flask-Testing
requests
pytest-flask
mock